refactor(viewer): apply viewer to Annotation Viewer Drawer
